requirements.txt
pyyaml==6.0.2
types-PyYAML==6.0.12.20240917
ruff==0.8.3
mypy==1.13.0
black==24.10.0
pyinstaller==6.15.0
pydantic==2.12.5
